+++ frontend/src/views/Tare/TareList/TareList.tsx
@@ -7,6 +7,7 @@ import CircularIndeterminate from "components/Loader/Loader";
 import { DefaultToolbar} from 'components';
 import {deleteTare, loadTare} from "redux/actions/tareAction";
 import {IStateInterface} from "redux/rootReducer";
+import { useConfirm } from "material-ui-confirm";
 
 
 const useStyles = makeStyles(theme => ({
@@ -22,6 +23,8 @@ const TareList = () => {
     const classes = useStyles();
     const history = useHistory();
     const dispatch = useDispatch();
+    const confirm = useConfirm();
+
 
     const tares = useSelector((state: IStateInterface) => state.tare.tareItems);
     const isLoading = useSelector((state: IStateInterface) => state.tare.isLoading);
@@ -37,9 +40,18 @@ const TareList = () => {
     }
 
     function onDeleteHandle() {
-        selected.forEach(async (item, i, selected) => {
-            dispatch(deleteTare(item))
-        });
+        confirm(
+            {
+                'title': 'Подтверждение',
+                description: `Удалить выбранные записи?.`,
+                confirmationText:'Да',
+                cancellationText: 'Нет'
+            }
+        ).then(() =>
+            selected.forEach(async (item, i, selected) => {
+                dispatch(deleteTare(item))
+            })
+        )
     }
 
     function onClickTableItem(tareId: number){
@@ -51,24 +63,25 @@ const TareList = () => {
         <div className={classes.root}>
         <DefaultToolbar
             className={''}
-    title={'Упаковочная тара продукции'}
-    newItemTitle={'Новая тара'}
-    newItemUrl={'/catalogs/tare/new'}
-    findCaption={'Поиск упаковочной тары'}
-    showDelete={true}
-    onFind={onFindProductHandler}
-    onDelete={onDeleteHandle}/>
-    <div className={classes.content}>
-        {isLoading ? <CircularIndeterminate/>
-                :
-                <TareTable
-                    tareItems={tares}
-                    onClickItem={onClickTableItem}
-                    className={''}
-                    onChangeSelected={setSelected}
-                />
-        }
-    </div>
+            title={'Упаковочная тара продукции'}
+            newItemTitle={'Новая тара'}
+            newItemUrl={'/catalogs/tare/new'}
+            findCaption={'Поиск упаковочной тары'}
+            showDelete={true}
+            onFind={onFindProductHandler}
+            onDelete={onDeleteHandle}
+        />
+            <div className={classes.content}>
+                {isLoading ? <CircularIndeterminate/>
+                        :
+                        <TareTable
+                            tareItems={tares}
+                            onClickItem={onClickTableItem}
+                            className={''}
+                            onChangeSelected={setSelected}
+                        />
+                }
+            </div>
     </div>
 );
 };
